analyze: refactor: move ExprRewrite into new mir_op submodule
